pyproject.toml
[tool.poetry]
name = "llm-rag-demo"
version = "0.1.0"
description = "A small project playing with large language models (LLMs) for retrieval augmented generation and a simple interface for both interactive queries and collecting data."
authors = ["Alex Baker"]
license = "Apache 2.0"
readme = "README.md"
packages = [{include = "llm_rag_demo"}]

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.29.0"
pandas = "^2.1.3"
langchain = "^0.0.344"
chromadb = "^0.4.18"
logging = "^0.4.9.6"
PyYAML = "^6.0.1"
pyarrow = "^14.0.1"
python-json-logger = "^2.0.7"
streamlit-chat = "^0.1.1"
st-star-rating = "^0.0.6"
sentence-transformers = "^2.2.2"
pypdf = "^3.17.4"
ipykernel = "^6.28.0"
pymupdf = "^1.23.8"
unstructured = "^0.11.6"
markdown = "^3.5.1"
docx2txt = "^0.8"
tiktoken = "^0.5.2"
instructorembedding = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
coverage = "^7.3.2"
pdoc = "^14.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
